Add anonymous user app authentication
This adds a way for an app to authenticate an anonymous
user to an app, while remaining anonymous in Teams. The
authentication page is created with the anonymous user id
as a parameter, then when auth succeeds, we return the
mapping of MSA <> user id. This is not secured as
such because of the implicit flow, but by replacing it with
a code exchange flow it would make it just fine.